pyproject.toml
[project]
name = "nn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium[box2d,classic-control]>=1.2.3",
    "numpy>=2.4.0",
    "pygame-ce>=2.5.6",
    "scikit-learn>=1.8.0",
    "stable-baselines3>=2.7.1",
    "swig>=4.4.1",
    "tabulate>=0.9.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
]
